requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
gunicorn
supabase
python-dotenv
